\name{archeofrag.gui}
\alias{archeofrag.gui}

\title{archeofrag.gui}
\description{Launch the archeofrag.gui application}

\usage{
archeofrag.gui()
}
\arguments{}
\details{

This function launches the `archeofrag.gui` application.
}
\value{
    Launch the `archeofrag.gui` Shiny application.
}
\author{Sebastien Plutniak <sebastien.plutniak at posteo.net>}

\seealso{
  \code{shiny}
  \code{archeofrag}
  \code{\link[shinythemes]{shinytheme}}
  \code{ggplot2}
}

\examples{
    \dontrun{
    # launching the app:
    archeofrag.gui()
    }
}
